[update] ディレクトリを変更
